+++ ipyrad/analysis/__init__.py
@@ -24,4 +24,4 @@ from .baba import Baba as baba
 from .bpp import Bpp as bpp
 from .sratools import SRA as sratools
 from .twiist import Twiist as twiist
-
+from .pca import PCA as pca