+++ ipyrad/assemble/consens_se.py
@@ -15,9 +15,9 @@ import datetime
 import pandas as pd
 import numpy as np
 import time
-import h5py
 import gzip
 import glob
+import io
 import os
 from ipyrad.assemble.jointestimate import recal_hidepth
 from util import TRANSFULL, progressbar, IPyradError, IPyradWarningExit, clustdealer, PRIORITY, MINOR
@@ -27,6 +27,10 @@ from collections import Counter
 import logging
 LOGGER = logging.getLogger(__name__)
 
+import warnings
+with warnings.catch_warnings():
+    warnings.filterwarnings("ignore", category=FutureWarning)
+    import h5py
 
 
 def get_binom(base1, base2, estE, estH):
@@ -942,7 +946,7 @@ def process_chunks(data, samples, lasyncs, lbview):
 
     ## write stats file
     data.stats_files.s5 = os.path.join(data.dirs.consens, 's5_consens_stats.txt')
-    with open(data.stats_files.s5, 'w') as out:
+    with io.open(data.stats_files.s5, 'w') as out:
         #out.write(data.stats_dfs.s5.to_string())
         data.stats_dfs.s5.to_string(
             buf=out,